# src/swagger/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Gym App API",
    "version": "1.0.0",
    "description": "API for Gym App"
  },
  "host": "localhost:3000",
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "basePath": "/",
  "schemes": [
    "http",
    "https"
  ],
  "paths": {
...
    "/exercise/all": {
      "get": {
        "tags": [
          "Exercise"
        ],
        "summary": "Get all exercises",
        "description": "Retrieves all exercises.",
        "responses": {
          "200": {
            "description": "Exercises found."
          },
          "404": {
            "description": "Exercises not found."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
    "/exercise/{uuid}": {
      "get": {
        "tags": [
          "Exercise"
        ],
        "summary": "Get an exercise by ID",
        "description": "Retrieves an exercise by its ID.",
        "parameters": [
          {
            "name": "uuid",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Exercise ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Exercise found."
          },
          "404": {
            "description": "Exercise not found."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
    "/exercise/name/{name}": {
      "get": {
        "tags": [
          "Exercise"
        ],
        "summary": "Get an exercise by name",
        "description": "Retrieves an exercise by its name.",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Exercise name"
          }
        ],
        "responses": {
          "200": {
            "description": "Exercise found."
          },
          "404": {
            "description": "Exercise not found."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
    "/exercise/muscle_group/{muscle_group}": {
      "get": {
        "tags": [
          "Exercise"
        ],
        "summary": "Get exercises by muscle group",
        "description": "Retrieves exercises by its muscle group.",
        "parameters": [
          {
            "name": "muscle_group",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Muscle group"
          }
        ],
        "responses": {
          "200": {
            "description": "Exercise found."
          },
          "404": {
            "description": "Exercise not found."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
    "/exercise/equipment/{equipment}": {
      "get": {
        "tags": [
          "Exercise"
        ],
        "summary": "Get exercises by equipment",
        "description": "Retrieves exercises by its equipment.",
        "parameters": [
          {
            "name": "equipment",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Equipment"
          }
        ],
        "responses": {
          "200": {
            "description": "Exercise found."
          },
          "404": {
            "description": "Exercise not found."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
    "/exercise/create": {
      "post": {
        "tags": [
          "Exercise"
        ],
        "summary": "Create a new exercise",
        "description": "Creates a new exercise.",
        "parameters": [
          {
            "name": "name",
            "description": "Exercise name",
            "required": true,
            "in": "query",
            "type": "string"
          },
          {
            "name": "muscle_group",
            "description": "Muscle group",
            "required": true,
            "in": "query",
            "type": "string"
          },
          {
            "name": "equipment",
            "description": "Equipment",
            "required": true,
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "201": {
            "description": "Exercise created."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
    "/exercise/update/{uuid}": {
      "put": {
        "tags": [
          "Exercise"
        ],
        "summary": "Update an exercise by ID",
        "description": "Updates an exercise by its ID.",
        "parameters": [
          {
            "name": "uuid",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Exercise ID"
          },
          {
            "name": "name",
            "description": "Exercise name",
            "in": "query",
            "type": "string"
          },
          {
            "name": "muscle_group",
            "description": "Muscle group",
            "in": "query",
            "type": "string"
          },
          {
            "name": "equipment",
            "description": "Equipment",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Exercise updated."
          },
          "404": {
            "description": "Exercise not found."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
...
    "/workout/delete/{uuid}/exercise/all": {
      "delete": {
        "tags": [
          "Workout"
        ],
        "summary": "Remove all exercises from workout by workout ID",
        "description": "Removes all exercises from a workout by its ID.",
        "parameters": [
          {
            "name": "uuid",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Workout ID"
          }
        ],
        "responses": {
          "204": {
            "description": "Exercises removed."
          },
          "404": {
            "description": "Workout not found."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
...
    "/workout/exercise/{uuid}": {
      "get": {
        "tags": [
          "Workout"
        ],
        "summary": "Get workouts by exercise ID",
        "description": "Retrieves workouts by exercise ID.",
        "parameters": [
          {
            "name": "uuid",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Exercise ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Workouts found."
          },
          "404": {
            "description": "Workouts not found."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
    "/workout/exercise/name/{name}": {
      "get": {
        "tags": [
          "Workout"
        ],
        "summary": "Get workouts by exercise name",
        "description": "Retrieves workouts by exercise name.",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Exercise name"
          }
        ],
        "responses": {
          "200": {
            "description": "Workouts found."
          },
          "404": {
            "description": "Workouts not found."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
...
  }
}